+++ src/main/java/org/example/models/Review.java
@@ -1,29 +1,29 @@
 package org.example.models;
 
 public class Review {
-    private final int reviewid;
-    private final int clientid;
+    private final Long reviewid;
+    private final Long clientid;
     private final String text;
-    private final int goodId;
+    private final Long goodId;
     private final int stars;
-    public Review(int reviewid, int clientid,int good_id, String text, int stars){
+    public Review(Long reviewid, Long clientid,Long good_id, String text, int stars){
         this.reviewid = reviewid;
         this.clientid = clientid;
         this.text = text;
         this.goodId = good_id;
         this.stars = stars;
     }
-    public int getGoodId(){
+    public Long getGoodId(){
         return goodId;
     }
-    public int getClientid(){
+    public Long getClientid(){
         return clientid;
     }
     public String getText(){
         return text;
     }
 
-    public int getReviewid(){
+    public Long getReviewid(){
         return reviewid;
     }
     public int getStars(){
